+++ quantities/bipm.hpp
@@ -1,6 +1,7 @@
 ﻿
 #pragma once
 
+#include "quantities/elementary_functions.hpp"
 #include "quantities/named_quantities.hpp"
 #include "quantities/numbers.hpp"
 #include "quantities/quantities.hpp"
