+++ quantities/quantities_body.hpp
@@ -1,13 +1,14 @@
 ﻿
 #pragma once
 
+#include "quantities/quantities.hpp"
+
 #include <cmath>
 #include <cstdio>
 #include <limits>
 #include <string>
 
 #include "base/macros.hpp"
-#include "quantities.hpp"
 
 namespace principia {
 namespace quantities {
@@ -172,6 +173,32 @@ struct ExponentiationGenerator<T, exponent, std::enable_if_t<(exponent == 1)>>{
   using Type = T;
 };
 
+template<int n, typename Q>
+struct NthRootGenerator<
+    n,
+    Q,
+    std::enable_if_t<(Q::Dimensions::Length % n) == 0 &&
+                     (Q::Dimensions::Mass % n) == 0 &&
+                     (Q::Dimensions::Time % n) == 0 &&
+                     (Q::Dimensions::Current % n) == 0 &&
+                     (Q::Dimensions::Temperature % n) == 0 &&
+                     (Q::Dimensions::Amount % n) == 0 &&
+                     (Q::Dimensions::LuminousIntensity % n) == 0 &&
+                     (Q::Dimensions::Angle % n) == 0>> {
+  enum {
+    Length            = Q::Dimensions::Length / n,
+    Mass              = Q::Dimensions::Mass / n,
+    Time              = Q::Dimensions::Time / n,
+    Current           = Q::Dimensions::Current / n,
+    Temperature       = Q::Dimensions::Temperature / n,
+    Amount            = Q::Dimensions::Amount / n,
+    LuminousIntensity = Q::Dimensions::LuminousIntensity / n,
+    Angle             = Q::Dimensions::Angle / n,
+  };
+  using Type = Quantity<Dimensions<Length, Mass, Time, Current, Temperature,
+                                   Amount, LuminousIntensity, Angle>>;
+};
+
 template<typename D>
 constexpr Quantity<D>::Quantity() : magnitude_(0) {}
 
@@ -286,8 +313,7 @@ constexpr Quantity<D> Quantity<D>::operator*(double const right) const {
 }
 
 template<typename LDimensions, typename RDimensions>
-FORCE_INLINE constexpr Product<Quantity<LDimensions>,
-                                       Quantity<RDimensions>>
+FORCE_INLINE constexpr Product<Quantity<LDimensions>, Quantity<RDimensions>>
 operator*(Quantity<LDimensions> const& left,
           Quantity<RDimensions> const& right) {
   return Product<Quantity<LDimensions>,
@@ -316,69 +342,6 @@ constexpr typename Quantity<RDimensions>::Inverse operator/(
   return typename Quantity<RDimensions>::Inverse(left / right.magnitude_);
 }
 
-template<int exponent>
-constexpr double Pow(double x) {
-  return std::pow(x, exponent);
-}
-
-// Static specializations for frequently-used exponents, so that this gets
-// turned into multiplications at compile time.
-
-template<>
-inline constexpr double Pow<-3>(double x) {
-  return 1 / (x * x * x);
-}
-
-template<>
-inline constexpr double Pow<-2>(double x) {
-  return 1 / (x * x);
-}
-
-template<>
-inline constexpr double Pow<-1>(double x) {
-  return 1 / x;
-}
-
-template<>
-inline constexpr double Pow<0>(double x) {
-  return 1;
-}
-
-template<>
-inline constexpr double Pow<1>(double x) {
-  return x;
-}
-
-template<>
-inline constexpr double Pow<2>(double x) {
-  return x * x;
-}
-
-template<>
-inline constexpr double Pow<3>(double x) {
-  return x * x * x;
-}
-
-template<int exponent, typename D>
-constexpr Exponentiation<Quantity<D>, exponent> Pow(
-    Quantity<D> const& x) {
-  return Exponentiation<Quantity<D>, exponent>(Pow<exponent>(x.magnitude_));
-}
-
-inline double Abs(double const x) {
-  return std::abs(x);
-}
-
-template<typename D>
-FORCE_INLINE Quantity<D> Abs(Quantity<D> const& quantity) {
-  return Quantity<D>(std::abs(quantity.magnitude_));
-}
-
-template<typename D>
-bool IsFinite(Quantity<D> const& x) {
-  return std::isfinite(x.magnitude_);
-}
-
 template<typename Q>
 constexpr Q Infinity() {
   return Q(std::numeric_limits<double>::infinity());
@@ -387,6 +350,11 @@ constexpr Q Infinity() {
 template<>
 constexpr double Infinity<double>() {
   return std::numeric_limits<double>::infinity();
+}
+
+template<typename D>
+bool IsFinite(Quantity<D> const& x) {
+  return std::isfinite(x.magnitude_);
 }
 
 template<typename Q>
